+++ src/loginServices.js
@@ -2,6 +2,7 @@ import localforage from 'localforage';
 import { config, translations } from 'stripes-config';
 import rtlDetect from 'rtl-detect';
 import moment from 'moment';
+import { loadDayJSLocale } from '@folio/stripes-components';
 
 import { discoverServices } from './discoverServices';
 import { resetStore } from './mainActions';
@@ -185,6 +186,11 @@ export function loadTranslations(store, locale, defaultTranslations = {}) {
       });
     });
   }
+
+  // load DayJS Locale. DayJS is expected to replace Moment as Stripes' date/time library.
+  // As with moment. loading the DayJS locale here passes it down to UI-modules so that they don't have to
+  // load the static locale data themselves.
+  loadDayJSLocale(locale);
 
   // Here we put additional condition because languages
   // like Japan we need to use like ja, but with numeric system